+++ CHANGELOG.md
@@ -8,7 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed 
 - BugFix: Small check in table class prevents crash on linux for tables with only one point
 - Changed ObjectUI Standard Behaviour for renaming to allow spaces - #109
+
+### Added
+- Shortcut settings for GTlab including the introduction of several new shortcuts and the option to maintain them in the preferences - #112
 - Versionnumber based on GTVersionnumber for tasks - #129
+###Removed
+- Removed DoubleMap container to be added to the post module - #116
+
 
 ## [1.7.2] - 2021-05-25
 ### Changed